wined3d: Remove brackets from shader_in_out varying
Influenced by the following glsl line:

Link info
---------
error: Varying (named shader_in_out.reg[0]) specified but not present in
the program object.

Includes a fixme in case it breaks something in the future.
I don't really know if this is right or not, but I do know this fixes
the GLSL link errors in OW without any visible negative side effects.

Signed-off-by: Zeke Sonxx <zeke@zekesonxx.com>